+++ sdk/instrumentation/library.go
@@ -22,4 +22,5 @@ For more information see
 package instrumentation // import "go.opentelemetry.io/otel/sdk/instrumentation"
 
 // Library represents the instrumentation library.
+// Deprecated: please use Scope instead.
 type Library = Scope